Address review feedback on per-step progress metrics
Changes based on reviewer comments:

1. Removed OpenTelemetry span events (addProgressEventToCurrentSpan):
   Per reviewer feedback, OTel is deprecating span events in favor of
   correlated logs. The per-step data is already logged at DEBUG level
   which is automatically correlated via trace/span IDs in the MDC.
   Reverted BatchJobOpenTelemetryUtils to its original state.

2. Kept the instance-level switch in updateCompletionStatus():
   The switch remains because it also captures error messages (first
   ERRORED message, last FAILED message) which are not tracked per-step.
   Added a comment explaining why the switch is intentionally kept.
   Used computeIfAbsent/merge for cleaner map operations.

3. Moved changelog to 8_12_0 folder (current dev version) and updated
   issue number to reference this PR (#7948) instead of unrelated #6424.

4. Restored the original TODO comment in calculateInstanceProgress():
   The status update still happens within the same function, so the
   TODO about separating it remains valid. Removed incorrect 'Phase'
   comments.

5. Renamed test to InstanceProgressTest with @Nested StepTracking class
   so future tests for other InstanceProgress behavior have a home.

6. Used assertThatThrownBy for the unmodifiable map test as suggested.
